chore: moved state folder, fix imports, moved components
