+++ the-easy-score/package.json
@@ -16,7 +16,6 @@
     "chart.js": "^2.9.4",
     "chartjs-plugin-datalabels": "^0.7.0",
     "chartjs-plugin-doughnutlabel": "^2.0.3",
-    "dotenv": "^8.2.0",
     "font-awesome": "^4.7.0",
     "fuse.js": "^6.4.6",
     "node-sass": "^4.14.1",
