+++ the-easy-score/src/components/Footer.js
