+++ the-easy-score/src/index.js
@@ -7,7 +7,7 @@ import { createStore, applyMiddleware, compose } from "redux";
 import { Provider } from "react-redux";
 import thunk from "redux-thunk";
 
-import reducer from "./reducers";
+import reducer from "./store/reducers";
 
 import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
 import "antd/dist/antd.css";
@@ -15,7 +15,6 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "./index.css";
 import "./custom.scss";
 import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 import ChartDataLabels from "chartjs-plugin-datalabels";
 import Chart from "chart.js";
 
@@ -42,8 +41,3 @@ ReactDOM.render(
   </Provider>,
   document.getElementById("root")
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
