+++ the-easy-score/src/screens/home/mobileSearch/MobileSearchForm.js
@@ -8,7 +8,7 @@ import {
   setCreditsFilterValue,
   setTimeFilterValue,
   resetFilterValues,
-} from "../../../actions/filtersActions";
+} from "../../../store/actions/filtersActions";
 
 import { FormControl, Button, InputGroup } from "react-bootstrap";
