+++ the-easy-score/src/screens/searchresults/Results.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { connect, useDispatch } from "react-redux";
-import { fetchResults, setCourses } from "../../actions/fetchDataActions";
-import { setNavStyle, setFooterStyle, setShowModal } from "../../actions/";
+import { fetchResults, setCourses } from "../../store/actions/fetchDataActions";
+import { setNavStyle, setFooterStyle, setShowModal } from "../../store/actions/";
 import { useParams } from "react-router-dom";
 
 import { filterToKeyword } from "./hooks/";
