+++ the-easy-score/src/static/url.js
@@ -1,0 +1,2 @@
+export const BASE_URL = process.env.REACT_APP_API_URL;
+// results&jsonquery=
