+++ the-easy-score/src/store/reducers/index.js
