--- the-easy-score/src/utils/api/getData.js
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-export const getData = () => {
-  return axios.create({
-    baseURL: `https://theeasyscore.com/results&jsonquery=`,
-  });
-};
